refactor: Session 구조 변경 및 동시성 컬렉션 반영
