+++ tomcat/src/main/java/org/apache/controller/LoginController.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import java.util.Set;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.catalina.Session;
+import org.apache.coyote.session.Session;
 import org.apache.controller.FileReader.FileReader;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.request.Request;
