+++ tomcat/src/main/java/org/apache/coyote/session/Session.java
@@ -1,12 +1,12 @@
-package org.apache.catalina;
+package org.apache.coyote.session;
 
-import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class Session {
 
     private final String id;
-    private final Map<String, Object> values = new HashMap<>();
+    private final Map<String, Object> values = new ConcurrentHashMap<>();
 
     public Session(String id) {
         this.id = id;
