+++ tomcat/src/test/java/org/apache/catalina/SessionManagerTest.java
@@ -3,6 +3,7 @@ package org.apache.catalina;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.UUID;
+import org.apache.coyote.session.SessionManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
