+++ tomcat/src/test/java/org/apache/coyote/request/RequestTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
-import org.apache.catalina.Session;
-import org.apache.catalina.SessionManager;
+import org.apache.coyote.session.Session;
+import org.apache.coyote.session.SessionManager;
 import org.apache.coyote.FixtureFactory;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
